Move person form to separate component
