+++ src/js/components/panes/PersonPane.jsx
@@ -1,9 +1,7 @@
 import React from 'react/addons';
 
 import PaneBase from './PaneBase';
-import Form from '../form/Form';
-import TextInput from '../form/TextInput';
-import SelectInput from '../form/SelectInput';
+import PersonForm from '../forms/PersonForm';
 
 
 export default class PersonPane extends PaneBase {
@@ -31,37 +29,10 @@ export default class PersonPane extends PaneBase {
 
     renderPaneContent(data) {
         if (data.person) {
-            var genderOptions = {
-                'f': 'Female',
-                'm': 'Male',
-                'o': 'Other'
-            };
-
             return (
-                <Form ref="personForm" onSubmit={ this.onSubmit.bind(this) }>
-                    <TextInput label="First name" name="first_name"
-                        initialValue={ data.person.first_name }/>
-                    <TextInput label="Last name" name="last_name"
-                        initialValue={ data.person.last_name }/>
-                    <SelectInput label="Gender" name="gender"
-                        options={ genderOptions }
-                        initialValue={ data.person.gender }/>
-                    <TextInput label="Email" name="email"
-                        initialValue={ data.person.email }/>
-                    <TextInput label="Phone number" name="phone"
-                        initialValue={ data.person.phone }/>
-
-                    <TextInput label="C/o address" name="co_address"
-                        initialValue={ data.person.co_address }/>
-                    <TextInput label="Street address" name="street_address"
-                        initialValue={ data.person.street_address }/>
-                    <TextInput label="Zip code" name="zip_code"
-                        initialValue={ data.person.zip_code }/>
-                    <TextInput label="City" name="city"
-                        initialValue={ data.person.city }/>
-
-                    <input name="submit" type="submit"/>
-                </Form>
+                <PersonForm ref="personForm"
+                    person={ data.person }
+                    onSubmit={ this.onSubmit.bind(this) }/>
             );
         }
         else {
